requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
supabase==2.0.0
reportlab==4.0.7
gunicorn==21.2.0
matplotlib==3.8.2
Pillow==10.1.0
